Use more specific types for -keyCode and -modifierFlags
It was initially suggested in #103 that we change -keyCode to NSInteger
and -modifierFlags to NSEventModifierFlags, but the changes were mixed
with other unrelated stuff and the discussion died out. But the main
point of the proposed patch was good, the suggested types make the code
work better in Swift, apart from other things.