Fix Deployment Target specified in podspec
Somewhere in the past we have bumped Deployment Target for the framework
to 10.10, but forgot to also bump the Deployment Target setting in the podspec,
leading to CocoaPods building the framework with different options. This lead
to mysterious warnings (#131) that were only visible in some CocoaPods builds.

This change fixes the CocoaPods Deployment Target to be the same as in the
Xcode project, ie. 10.10.